Move Transaction type to primitives
Needs no explanation - lets go!
